parse_arg and load_config outsourced to utils; easily reusable in e.g. condorJobs_reco
